feat(desktop): per-domain agent/session/event bridges + tauri-specta event types
- Add `tauri_specta::Event` derive to all 5 event structs in events.rs
- Register events via `.events(collect_events![...])` in make_builder() so they appear in generated.ts
- Change UsageEvent token fields from u64 → u32 (prevents specta BigInt export error; saturating cast in agent_service.rs)
- Create desktop/src/lib/bridge/agent.ts, session.ts, events.ts with typed wrappers over client.ts::call and @tauri-apps/api/event::listen
- Update bridge/index.ts with agentBridge, sessionBridge, events, and clean type aliases
- Regenerate generated.ts with all 5 event type shapes + SessionMessage_Serialize alias

+++ desktop/src-tauri/src/events.rs
@@ -4,13 +4,13 @@
 use serde::Serialize;
 use tauri::{AppHandle, Emitter};
 
-#[derive(Debug, Clone, Serialize, PartialEq, specta::Type)]
+#[derive(Debug, Clone, Serialize, PartialEq, specta::Type, tauri_specta::Event)]
 pub struct TextDeltaEvent {
     pub session_id: String,
     pub text: String,
 }
 
-#[derive(Debug, Clone, Serialize, PartialEq, specta::Type)]
+#[derive(Debug, Clone, Serialize, PartialEq, specta::Type, tauri_specta::Event)]
 pub struct ToolCallDeltaEvent {
     pub session_id: String,
     pub index: u32,
@@ -22,24 +22,26 @@ pub struct ToolCallDeltaEvent {
     pub call_id: Option<String>,
 }
 
-#[derive(Debug, Clone, Serialize, PartialEq, specta::Type)]
+#[derive(Debug, Clone, Serialize, PartialEq, specta::Type, tauri_specta::Event)]
 pub struct UsageEvent {
     pub session_id: String,
-    pub prompt_tokens: u64,
-    pub completion_tokens: u64,
-    pub total_tokens: u64,
+    /// Token counts are capped at u32 for frontend JSON-number safety (u64 exceeds
+    /// JS Number precision; LLM responses comfortably fit in ~4 B tokens).
+    pub prompt_tokens: u32,
+    pub completion_tokens: u32,
+    pub total_tokens: u32,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub estimated_cost: Option<f64>,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, specta::Type)]
+#[derive(Debug, Clone, PartialEq, Serialize, specta::Type, tauri_specta::Event)]
 pub struct DoneEvent {
     pub session_id: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub reason: Option<String>,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, specta::Type)]
+#[derive(Debug, Clone, PartialEq, Serialize, specta::Type, tauri_specta::Event)]
 pub struct ErrorEvent {
     pub session_id: String,
     pub message: String,
